requirements/base.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot

awslambdaric
click
boto3
botocore
black